Fixed MP5 iron sight bug and added realistic weapon weight
There was an ironsight bug with the unsuppressed MP5 where only the front sight was visible. Copying and pasting the Ironsight offsets from the suppressed MP5 fixes this issue.
I also made the FAL use the same ammo values as the SCAR since they share the 7.62x51 round. This does reduce the FAL mag capacity from 30 to 20, though.